near-rubric-mcp/requirements.txt
PyYAML==6.0
ujson==5.8.0
asyncio==3.4.3 